Split app into a menu bar app, and a 'helper' GUI app
Splitting the app into two processes, one that handles purely the menu
bar icon, and another 'helper' app that handles GUI windows, means that
several issues can be avoided.

Specifically, this design makes it easier to show the dock icon, and the
main menu bar for the app, when a GUI window is visible. The menu bar
and the dock icon can then be removed be quitting the helper app.

An alternative is to use -[NSApplication setActivationPolicy:] but
this approach seemed to cause some problems. The menu bar would freeze
sometimes until the app gained and lost focus. In addition, the
documentation suggests that this method may not be fully supported.

The helper app and the menu bar app communicate via TCP sockets using
the CFSocket and NSStream APIs.